// client/src/app/phonelisting.ts
import { User } from './user';
import { Review } from "./review";


export interface PhoneListing {
    _id: string,
    title: string;
    brand: string;
    stock: number;
    seller: User;
    price: number;
    reviews: Review[];
    disabled: boolean;
    avgRating?: number;
    image?: string;
}